Replace ccleaner with sdelete from microsoft
For nulling the free space cc cleaner is recommended in the documentation. As this software can contain adware and has to be installed first, I'd recommend SDelete from Sysinternals (https://technet.microsoft.com/en-us/library/2009.08.utilityspotlight.aspx). 
Advantage: It's an official tool from Microsoft and it's a command line utility